+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added similarity score based on comparing parent masses [#79](https://github.com/matchms/matchms/pull/79)
+- Added a method for instantiating a spectrum from the metabolomics USI 
 
 ## [0.4.0] - 2020-06-11